+++ lib/core/version/app_release.dart
@@ -7,8 +7,8 @@ import '../../l10n/app_localizations.dart';
 class AppRelease {
   const AppRelease._();
 
-  static const version = '0.1.8';
-  static const buildNumber = 9;
+  static const version = '0.1.9';
+  static const buildNumber = 10;
   static const fullVersion = '$version+$buildNumber';
   static const _lastSeenPreference = 'last_seen_release';
 }